+++ Tests/PositronicKitTests/Services/GroupedInitToolApprovalPolicyWiringTests.swift
@@ -68,7 +68,6 @@ struct GroupedInitToolApprovalPolicyWiringTests {
         let persistence = PKRuntime.PersistenceConfiguration(
             runtimeRepository: mockPersistence,
             workspacePersistence: mockPersistence,
-            toolPersistence: mockPersistence,
             agentStore: mockPersistence,
             requestOriginStore: mockPersistence
         )
@@ -98,7 +97,6 @@ struct GroupedInitToolApprovalPolicyWiringTests {
         let persistence = PKRuntime.PersistenceConfiguration(
             runtimeRepository: mockPersistence,
             workspacePersistence: mockPersistence,
-            toolPersistence: mockPersistence,
             agentStore: mockPersistence,
             requestOriginStore: mockPersistence
         )